+++ crates/maelstrom-worker/src/config.rs
@@ -13,24 +13,28 @@ pub struct Config {
     pub broker: BrokerAddr,
 
     /// The number of job slots available.
-    #[config(short = 'S', value_name = "N", default = "Slots::default()")]
+    #[config(value_name = "N", default = "Slots::default()")]
     pub slots: Slots,
 
     /// The directory to use for the cache.
     #[config(
-        short = 'r',
         value_name = "PATH",
-        default = "|bd: &BaseDirectories| bd.get_cache_home().into_os_string().into_string().unwrap()"
+        default = r#"|bd: &BaseDirectories| {
+            bd.get_cache_home()
+                .into_os_string()
+                .into_string()
+                .unwrap()
+        }"#
     )]
     pub cache_root: RootBuf<WorkerCacheDir>,
 
     /// The target amount of disk space to use for the cache. This bound won't be followed
     /// strictly, so it's best to be conservative. SI and binary suffixes are supported.
-    #[config(short = 's', value_name = "BYTES", default = "CacheSize::default()")]
+    #[config(value_name = "BYTES", default = "CacheSize::default()")]
     pub cache_size: CacheSize,
 
     /// The maximum amount of bytes to return inline for captured stdout and stderr.
-    #[config(short = 'i', value_name = "BYTES", default = "InlineLimit::default()")]
+    #[config(value_name = "BYTES", default = "InlineLimit::default()")]
     pub inline_limit: InlineLimit,
 
     /// Minimum log level to output.